% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clientprofile.R
\name{agent_filters_UI}
\alias{agent_filters_UI}
\title{Agent Filters UI}
\usage{
agent_filters_UI(ns, auth_res, conditions)
}
\arguments{
\item{ns}{The namespace for the Shiny inputs.}

\item{agent}{The agent object.}
}
\value{
A div element containing the agent filters user interface.
}
\description{
This function creates a user interface for agent filters.
}
