% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregations.R
\docType{data}
\name{aggregations}
\alias{aggregations}
\title{aggregations}
\format{
A data frame with 148 rows and 59 variables
\itemize{
\item  id   character
\item  TA   character
\item  CA   character
\item  CIA   character
\item  COA   character
\item  NCA   character
\item  NCOA   character
\item  NCIA   character
\item  TL   character
\item  CL   character
\item  COL   character
\item  CFL   character
\item  DP   character
\item  NCL   character
\item  NCFL   character
\item  NCOL   character
\item  SE   character
\item  IC   character
\item  IBD   character
\item  OWI   character
\item  PC   character
\item  WC   character
\item  CE   character
\item  NFPC   character
\item  NFWC   character
\item  CASH   character
\item  QA   character
\item  AR   character
\item  INV   character
\item  AP   character
\item  UDR   character
\item  REV   character
\item  COS   character
\item  GM   character
\item  SGNA   character
\item  EBITDA   character
\item  DEPR   character
\item  OI   character
\item  NOGL   character
\item  EBIT   character
\item  INT   character
\item  EBT   character
\item  TAX   character
\item  NI   character
\item  CFOA   character
\item  CFOA.NI   character
\item  CFOA.RECON   character
\item  CFOA.OP   character
\item  TCFOA   character
\item  CFIA   character
\item  CFIA.OP   character
\item  CFIA.INV   character
\item  TCFIA   character
\item  CFFA   character
\item  CFFA.DBT   character
\item  CFFA.EQT   character
\item  TCFFA   character
\item  OACF   character
\item  CASHVAR   character
}
}
\usage{
data(aggregations)
}
\description{
aggregations
}
\keyword{datasets}
