% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_fs_labels.R
\name{classify_fs_labels}
\alias{classify_fs_labels}
\title{Classify labels in financial statements}
\usage{
classify_fs_labels(label, section)
}
\arguments{
\item{label}{Character. Label to classify.}

\item{section}{Character. Preliminary classification of the label.}
}
\value{
Character. ID for the account.
}
\description{
Assign an account ID to financial statements accounts based on a preliminary classification and the terms appearing in the label.
}
\author{
Nicolas Mangin
}
