% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_fs_dates.R
\name{clean_fs_dates}
\alias{clean_fs_dates}
\title{Correct dates in financial statements}
\usage{
clean_fs_dates(finstat, filings)
}
\arguments{
\item{finstat}{Tibble. Statements gathered from csv files.}

\item{filings}{Tibble. List of SEC filings.}
}
\value{
Tibble. Financial statements with clean dates
}
\description{
Function correcting dates and reporting dates to make them consistent with SEC filings' information
}
\author{
Nicolas Mangin
}
