% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect}
\alias{collect}
\title{Add an item to a collectr collection}
\usage{
collect(item, ..., collection = NULL)
}
\arguments{
\item{item}{An object to be added. If not a \code{collectr.item}, it will be wrapped as one.}

\item{...}{Named options for the new item's specific options.}

\item{collection}{A \code{collectr} object to add the item to. If NULL, a new \code{collectr.item} is created.}
}
\value{
A modified \code{collectr} object with the new item included.
}
\description{
Adds an item to a collectr collection, either as a \code{collectr.item} object or a generic R object.
}
