% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_settings.R
\name{collect_contacts}
\alias{collect_contacts}
\title{Collect formatted contact information from a list}
\usage{
collect_contacts(input, id)
}
\arguments{
\item{input}{A list containing contact information.}

\item{id}{An integer specifying the number of contacts in the list.}
}
\value{
A data frame containing the formatted contact information for all contacts in the list.
}
\description{
This function takes a list of contact information and the number of contacts in the list,
and returns a data frame containing the formatted contact information for all contacts in the list.
}
