% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectr.R
\name{collectr}
\alias{collectr}
\title{Collectr wrapper}
\usage{
collectr(family = NULL, ...)
}
\arguments{
\item{family}{character(1), must be one of collectr_families() or NULL}

\item{...}{options to be substituted on top of the collectr_default_options(family)}
}
\value{
structure of class collectr and sub-classes by family
}
\description{
Collectr wrapper
}
