% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectr.R
\name{collectr_default_options}
\alias{collectr_default_options}
\title{Default options for collectr collections}
\usage{
collectr_default_options(family = NULL)
}
\arguments{
\item{family}{character(1), must be one of \code{collectr_families()}, or NULL to return general defaults.}
}
\value{
A named list of default options, including \code{"save_path"} and \code{"file"}. If a family is provided, additional parameters may be included.
}
\description{
Returns default options for collectr based on a specified family.
}
