% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectr_families.R
\name{collectr_families}
\alias{collectr_families}
\title{Supported families of collectr collections}
\usage{
collectr_families(quite = FALSE)
}
\arguments{
\item{quiet}{logical(1), defaults to FALSE. If TRUE, suppresses the message describing available families.}
}
\value{
character vector of supported collectr families.
}
\description{
Returns a character vector of available collectr families that can be used to configure collections.
}
