% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colrect.R
\name{colrect}
\alias{colrect}
\title{Collect clustered data on multiple sheets into rectangle dataframe}
\usage{
colrect(
  path,
  sheet_regex,
  row = NULL,
  col = NULL,
  regex = NULL,
  offset = c(0, 0),
  ends = NULL,
  info = NULL
)
}
\arguments{
\item{path}{Path to the xls/xlsx file.}

\item{sheet_regex}{regex to search key}

\item{row}{Row position of df where the keyword appears}

\item{col}{Column position of df where the keyword appears}

\item{regex}{Regular expression to match keywords}

\item{offset}{The offset (\code{c(row, pos})) of the cluster topleft from
the coordination of keyword}

\item{ends}{List of regex to locate row- and column- ends of each cluster
Form should be like \code{ends = list(row = "2019", col = "[Dd]ecember$")}.
Regex \code{row = } must specify the end of 'left most' columnn of df,
not that of the column with key matched by \code{regex}}

\item{info}{Parameters to make key:value list such as
\describe{
\item{key_offset}{Offset \code{c(row, col)} of \code{key} topleft
from df topleft. If \code{NULL}, automatically set to \code{keyn}}
\item{key_dim}{Dimension \code{c(row, col)} of \code{key}}
\item{value_offset}{Offset \code{c(row, col)} of \code{value} topleft from
df topleft}
\item{value_dim}{Dimension \code{c(row, col)} of \code{value}}
}}
}
\description{
Collect clustered data on multiple sheets into rectangle dataframe
}
