% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_cik.R
\name{complete_cik}
\alias{complete_cik}
\title{Standardize cik format}
\usage{
complete_cik(cik)
}
\arguments{
\item{cik}{Character. Short CIK code}
}
\value{
Character. Long CIK code
}
\description{
add CIK and ensures a 10 digits cik code.
}
\author{
Nicolas Mangin
}
