% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cost_of_capital.R
\name{compute_cost_of_capital}
\alias{compute_cost_of_capital}
\title{Compute cost of capital}
\usage{
compute_cost_of_capital(tickers = NA, market = "^GSPC")
}
\arguments{
\item{tickers}{Character vector. Tickers of the companies for which the cost of capital should be computed.}

\item{market}{Character. Ticker of the market used as reference.}
}
\value{
Tibble. cik, id, date, value
}
\description{
Compute various components of the cost of capital.
}
\author{
Nicolas Mangin
}
