% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_financial_data.R
\name{compute_financial_data}
\alias{compute_financial_data}
\title{Create financial data set}
\usage{
compute_financial_data(dataset = NA, type = "observation")
}
\arguments{
\item{dataset}{Character or tibble. Either the cik code of the company or a simulated dataset.}

\item{type}{Character. Whether the dataset should retrieved from the cik code ("observation") or given ("simulation")}
}
\value{
Tibble with financial data
}
\description{
Create a complete set of statements, aggregates and ratios for the selected company in the chosen language.
}
\author{
Nicolas Mangin
}
