% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_share_yield.R
\name{compute_share_yield}
\alias{compute_share_yield}
\title{Compute Share yield}
\usage{
compute_share_yield(dataset = NA)
}
\arguments{
\item{dataset}{Character. CIK code of the company.}
}
\value{
Tibble with financial data
}
\description{
Compute the share yield or total share return (including both dividends and change in value)
}
\author{
Nicolas Mangin
}
