% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corporations.R
\docType{data}
\name{corporations}
\alias{corporations}
\title{corporations}
\format{
A data frame with 503 rows and 10 variables
\itemize{
\item  cik   character
\item  ticker   character
\item  name   character
\item  sector   character
\item  industry   character
\item  headquarter   character
\item  inSNP   character
\item  founded   character
\item  url_nyse   character
\item  url_wikipedia   character
}
}
\usage{
data(corporations)
}
\description{
corporations
}
\keyword{datasets}
