% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_interest.R
\name{correct_interest}
\alias{correct_interest}
\title{Correct interest}
\usage{
correct_interest(x)
}
\arguments{
\item{x}{Tibble. Statements filtered on INT and PAIDINT.}
}
\value{
Character. ID for the account.
}
\description{
Increase the interest expense and acknowledge a financial gain to account for the gap betweon interest expense and interest paid.
}
\author{
Nicolas Mangin
}
