% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_of_capital.R
\docType{data}
\name{cost_of_capital}
\alias{cost_of_capital}
\title{cost_of_capital}
\format{
A data frame with 1515 rows and 4 variables
\itemize{
\item  cik   character
\item  id   character
\item  date   character
\item  value   character
}
}
\usage{
data(cost_of_capital)
}
\description{
cost_of_capital
}
\keyword{datasets}
