% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\name{crop}
\alias{crop}
\title{Crop data frame at specific keyword}
\usage{
crop(df, direction = NULL, pos = NULL, regex = NULL, use_after = FALSE)
}
\arguments{
\item{df}{Data frame to be processed}

\item{direction}{Direction of the vector to be pulled from df
#' \describe{
\item{"h"}{specific row of the df will be returned as avector}
\item{"v"}{specific column of the df will be returned as a vector}
}}

\item{pos}{Row- or column position to be searched}

\item{regex}{regex to match keyword}

\item{use_after}{If TRUE, use part after match}
}
\description{
Crop data frame at specific keyword
}
