% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_customfunctions.R
\name{custom_db_actions}
\alias{custom_db_actions}
\title{Execute an action in the database and show a toast message}
\usage{
custom_db_actions(action, successmessage, toast = T, ...)
}
\arguments{
\item{action}{A database action to be executed.}

\item{successmessage}{A message to display if the action is successful.}
}
\description{
This function executes a specified action in the database and shows a toast message
indicating whether the action was successful or not.
}
