% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db_actions.R
\name{delete_clause}
\alias{delete_clause}
\title{Generate an SQL insert statement for a table
Generate a DELETE SQL clause}
\usage{
delete_clause(con, table, where, is, all = FALSE, exec = TRUE)
}
\arguments{
\item{con}{The database connection object.}

\item{table}{The name of the table to delete rows from.}

\item{where}{The column name used to specify the WHERE condition.}

\item{is}{The value to match in the WHERE condition.}

\item{exec}{Logical value indicating whether to execute the generated SQL or return it as a string.}
}
\value{
If \code{exec = TRUE}, this function returns the result of the executed SQL DELETE statement. If \code{exec = FALSE}, this function returns a character string representing the SQL DELETE statement.
}
\description{
This function generates a SQL DELETE clause for a given table and WHERE condition, which deletes rows from the table that match the condition.
}
