% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_customfunctions.R
\name{delete_confirm}
\alias{delete_confirm}
\title{Execute an action in the database and show a toast message}
\usage{
delete_confirm(text, delete_id, ns)
}
\arguments{
\item{text}{The dismisbutton text}

\item{delete_id}{The unique delete id}

\item{successmessage}{A message to display if the action is successful.}
}
\description{
This function executes a specified action in the database and shows a toast message
indicating whether the action was successful or not.
}
