% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_fs_pattern.R
\name{detect_fs_pattern}
\alias{detect_fs_pattern}
\title{Detect pattern}
\usage{
detect_fs_pattern(x, y)
}
\arguments{
\item{x}{Character vector.}

\item{y}{Character. A pattern to detect.}
}
\value{
Logical. Whether the character vector contains a specific pattern.
}
\description{
Check whether a character vector contains a specific pattern.
}
\author{
Nicolas Mangin
}
