% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_info_in_col.R
\name{detect_info_in_col}
\alias{detect_info_in_col}
\title{FDetect information}
\usage{
detect_info_in_col(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Logical. Whether the character vector contains some pieces of information.
}
\description{
Check whether a column contains some pieces of information.
}
\author{
Nicolas Mangin
}
