% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{.collect}
\alias{.collect}
\title{Internal function to add an item to a collectr collection}
\usage{
.collect(item, collection)
}
\arguments{
\item{item}{A \code{collectr.item} object}

\item{collection}{A \code{collectr} object to which the item is added.}
}
\value{
A modified collectr object with the new item.
}
\description{
Adds an item to an existing collectr collection. Not intended for direct use.
}
\keyword{internal}
