% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_settings.R
\name{edit_group_permission}
\alias{edit_group_permission}
\title{Edit group permissions}
\usage{
edit_group_permission(groups, selected_row, permissions_selected, group_name)
}
\arguments{
\item{selected_row}{A character vector with the selected row to edit.}

\item{permissions_selected}{A character vector with the selected permissions.}

\item{group_name}{A character vector with the name of the group to edit.}
}
\value{
Nothing is returned. The function modifies the database table 't_GroupPermissions'.
}
\description{
This function edits the permissions of a given group based on the selected row and the list of permissions selected.
}
