% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_settings.R
\name{edit_permission_ui}
\alias{edit_permission_ui}
\title{Edit UI Function}
\usage{
edit_permission_ui(what, ns, data, title = "Edit Permission Details")
}
\arguments{
\item{what}{A character string indicating the type of data to be edited (permissions or users).}

\item{ns}{A character string with the namespace for Shiny input and output objects.}

\item{data}{A data frame with the data to be edited.}
}
\value{
A modal dialog with input fields to edit the data of a given \code{what} (permissions or users).
}
\description{
This function generates a modal dialog with input fields to edit the data of a given \code{what} (permissions or users) and returns it.
}
