% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_settings.R
\name{enc_dec}
\alias{enc_dec}
\title{Encrypt or decrypt a password}
\usage{
enc_dec(passwd, enc = TRUE)
}
\arguments{
\item{passwd}{A character string containing the password to be encrypted or decrypted.}

\item{enc}{A logical value indicating whether to encrypt or decrypt the password.
The default value is \code{TRUE}.}
}
\value{
A character string containing the encrypted or decrypted password.
}
\description{
This function encrypts or decrypts a given password using the \code{safer} package
and a key obtained from the \code{get_golem_config} function.
}
