% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_allocationclass.R
\name{external_team_outsource}
\alias{external_team_outsource}
\title{Perform account allocation to external team based on outsource method.}
\usage{
external_team_outsource(
  id_col,
  dt,
  outsource_method,
  top_n = 3,
  manual_agents = NULL,
  manual_ratio = NULL,
  test = TRUE
)
}
\arguments{
\item{dt}{The allocation data frame.}

\item{outsource_method}{The outsource method, either "equal" or "ratio".}

\item{top_n}{The number of top-performing agents to consider.}

\item{manual_agents}{A list of manual agents to consider for each channel (optional).}

\item{manual_ratio}{A list of manual allocation ratios for each channel (optional).}

\item{test}{Logical indicating whether to run a test mode (default is TRUE).}
}
\value{
The allocation data frame.
}
\description{
Perform account allocation to external team based on outsource method.
}
