% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_agentPerformanceDashboard.R
\name{feedbac_data}
\alias{feedbac_data}
\title{Fetch Feedback Data}
\usage{
feedbac_data(channel_selected, debt_collector, year_val, month_val, team_val)
}
\arguments{
\item{channel_selected}{The selected channel for feedback data. Use "All" for all channels.}

\item{debt_collector}{The name of the debt collector.}

\item{year_val}{The year for filtering feedback data.}

\item{month_val}{The month for filtering feedback data.}
}
\value{
A data frame with consolidated feedback information including count, outsourced amount, amount recovered, and recovery rate.
}
\description{
Retrieves feedback data based on the specified parameters.
}
