% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filings.R
\docType{data}
\name{filings}
\alias{filings}
\title{filings}
\format{
A data frame with 579 rows and 5 variables
\itemize{
\item  cik   character
\item  form   character
\item  filingDate   character
\item  reportDate   character
\item  link   character
}
}
\usage{
data(filings)
}
\description{
filings
}
\keyword{datasets}
