% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_submission.R
\name{flatten_submission}
\alias{flatten_submission}
\title{Flatten submission}
\usage{
flatten_submission(json_file)
}
\arguments{
\item{json_file}{Character. Path to the local json file.}
}
\value{
Tibble. Submissions information.
}
\description{
Transform a SEC JSON file into a tibble with the relevant information.
}
\author{
Nicolas Mangin
}
