% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_statement.R
\name{format_statement}
\alias{format_statement}
\title{Format a statement}
\usage{
format_statement(statement)
}
\arguments{
\item{statement}{List. 3 Financial statements: BS, IS, CFS.}
}
\value{
Tibble. Statements.
}
\description{
Format and combine the selected and retrieved tables for each financial statement is a list.
}
\author{
Nicolas Mangin
}
