% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_allocationclass.R
\name{get_allocation_file}
\alias{get_allocation_file}
\title{Get the allocation file for debt collectors.}
\usage{
get_allocation_file(data, report_date)
}
\value{
A list containing the "carried_foward" and "new_cases" data frames.
}
\description{
Get the allocation file for debt collectors.
}
