% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_esg_sustainalytics.R
\name{get_esg_sustainalytics}
\alias{get_esg_sustainalytics}
\title{Retreive ESG data}
\usage{
get_esg_sustainalytics(companies = NA)
}
\arguments{
\item{companies}{Tibble List of companies with one column called "ticker"}
}
\value{
Tibble with ESG scores and areas of involvement
}
\description{
Retreive Sustainalytics' ESG scores from Yahoo Finance.
}
\author{
Nicolas Mangin
}
