% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filings.R
\name{get_filings}
\alias{get_filings}
\title{Classify labels in financial statements}
\usage{
get_filings(folder)
}
\arguments{
\item{folder}{Character. path to the folder containing filings lists.}
}
\value{
Character. ID for the account.
}
\description{
Assign an account ID to financial statements accounts based on a preliminary classification and the terms appearing in the label.
}
\author{
Nicolas Mangin
}
