% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fs_files.R
\name{get_fs_files}
\alias{get_fs_files}
\title{Retrieve the content of csv files.}
\usage{
get_fs_files(folder)
}
\arguments{
\item{folder}{Character. Path to csv files.}
}
\value{
Tibble. Content of csv files.
}
\description{
Retrieve the content of csv files.
}
\author{
Nicolas Mangin
}
