% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_allocationclass.R
\name{get_latest_month_perf}
\alias{get_latest_month_perf}
\title{Get the latest month's performance for external debt collectors.}
\usage{
get_latest_month_perf(top_n)
}
\arguments{
\item{top_n}{The number of top-performing agents to consider.}
}
\value{
A list of data frames containing allocation ratios for each channel.
}
\description{
Get the latest month's performance for external debt collectors.
}
