% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_market_data.R
\name{get_market_data}
\alias{get_market_data}
\title{Collect market data}
\usage{
get_market_data(
  symbol = "AAPL",
  fmpkey = NA,
  start = "2019-01-01",
  end = "2024-12-31"
)
}
\arguments{
\item{symbol}{Character. Ticker of the company.}

\item{fmpkey}{Character. API key for FMP.}

\item{start}{Character. Starting date.}

\item{end}{Character. Ending date.}
}
\value{
Tibble.
}
\description{
Collect stock prices, dividends, market capitalization and number of common outstanding shares for a company.
}
\author{
Nicolas Mangin
}
