% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_market_indexes.R
\name{get_market_indexes}
\alias{get_market_indexes}
\title{Get indexes}
\usage{
get_market_indexes(start = "2019-01-01", end = "2025-01-01")
}
\arguments{
\item{start}{Character. Starting date}

\item{end}{Character. Ending date}
}
\value{
Tibble
}
\description{
Get index levels for the NYSE, NASDAQ, and SP500.
}
\author{
Nicolas Mangin
}
