% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_settings.R
\name{get_new_group_permissions}
\alias{get_new_group_permissions}
\title{Get new group permissions}
\usage{
get_new_group_permissions(selected_permissions, group_name)
}
\arguments{
\item{selected_permissions}{A character vector with the selected permissions.}

\item{group_name}{A character vector with the name of the group.}
}
\value{
A data frame with the new permissions for the given group.
}
\description{
This function returns a data frame with the new permissions for a given group.
}
