% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_periodic_stock_prices.R
\name{get_periodic_stock_prices}
\alias{get_periodic_stock_prices}
\title{Retrieve stock prices}
\usage{
get_periodic_stock_prices(tickers, mindate, maxdate, period = "monthly")
}
\arguments{
\item{tickers}{Character vector. List of tickers.}

\item{mindate}{Character. Beginning date ("2020-01-01)}

\item{maxdate}{Character. Ending date ("2022-12-01)}

\item{period}{Character. "monthly" or "daily".}
}
\value{
Tibble. ticker, date, adjusted price.
}
\description{
Retrieve stock prices for a vector of corporations' tickers.
}
\author{
Nicolas Mangin
}
