% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_allocationclass.R
\name{get_scoring_files}
\alias{get_scoring_files}
\title{Get the files required for allocation.}
\usage{
get_scoring_files(report_date)
}
\value{
A list containing the "outstandingfile" and "loanbook" data frames.
}
\description{
Get the files required for allocation.
}
