% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_share_number.R
\name{get_share_number}
\alias{get_share_number}
\title{Get share number}
\usage{
get_share_number(html_file)
}
\arguments{
\item{html_file}{Character. Path to the local HTML file.}
}
\value{
Integer. Number of shares.
}
\description{
Retrieve the number of shares from a HTML annual report.
}
\author{
Nicolas Mangin
}
