% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sp500.R
\name{get_sp500}
\alias{get_sp500}
\title{Retrieve S&P500 companies}
\usage{
get_sp500()
}
\value{
Tibble with the list of S&P 500 companies.
}
\description{
Retrieve S&P500 companies from Wikipedia.
}
\author{
Nicolas Mangin
}
