% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables_from_html.R
\name{get_tables_from_html}
\alias{get_tables_from_html}
\title{Retrieve HTML tables}
\usage{
get_tables_from_html(file)
}
\arguments{
\item{file}{Character. Path to the local copy of the HTML file.}
}
\value{
List of tables.
}
\description{
Retrieve all the tables from a HTML 10-K annual report.
}
\author{
Nicolas Mangin
}
