% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_treasury_yield.R
\name{get_treasury_yield}
\alias{get_treasury_yield}
\title{Get US risk-gree rate}
\usage{
get_treasury_yield(start = "2019-01-01", end = "2025-01-01")
}
\arguments{
\item{start}{Character. Starting date}

\item{end}{Character. Ending date}
}
\value{
Tibble
}
\description{
Collect US treasury yield data.
}
\author{
Nicolas Mangin
}
