% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_us_lt_rates.R
\name{get_us_lt_rates}
\alias{get_us_lt_rates}
\title{Retrieve US long-term interest rates}
\usage{
get_us_lt_rates()
}
\value{
Tibble. period, value
}
\description{
Retrieve US long-term interest rates
}
\author{
Nicolas Mangin
}
