% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectr_methods.R
\name{\%++\%}
\alias{\%++\%}
\title{Adds and item to a collection in the re-assign manner.}
\usage{
lhs \%++\% rhs
}
\arguments{
\item{lhs}{A \code{collectr} object.}

\item{rhs}{A \code{collectr.item} object.}

\item{lhs_name}{NULL or character(1), character is only propagated by a method call with shortcut = TRUE}

\item{shortcut}{logical(1) defaults to FALSE only set to TRUE by S3 method call}
}
\value{
collectr object, invisible, extended collectr collection.
}
\description{
Adds and item to a collection in the re-assign manner.
}
