% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_settings.R
\name{group_settings}
\alias{group_settings}
\title{Show modal for editing group settings}
\usage{
group_settings(
  ns,
  fun,
  group_auth,
  user_auth,
  title,
  value = NULL,
  selected_perms = NULL
)
}
\arguments{
\item{fun}{The function to be executed when the Save button is clicked.}

\item{group_auth}{A character vector of group-level authorization settings.}

\item{user_auth}{A character vector of user-level authorization settings.}
}
\value{
A modal dialog containing input fields for editing group settings.
}
\description{
This function displays a modal dialog for editing the settings of a group.
}
