% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_tables.R
\name{identify_tables}
\alias{identify_tables}
\title{Find statements positions}
\usage{
identify_tables(tables, all_patterns, remove = NA, minscore = 3, window = 2)
}
\arguments{
\item{tables}{List. all tables from a 10-K annual report.}

\item{all_patterns}{List of character vectors. Patterns identifying the statements.}

\item{remove}{Numeric vector. Position of the tables to be removed.}

\item{minscore}{Integer. Minimum number of pattern occurrences to keep a table as a candidate for a statement.}

\item{window}{Integer. To keep candidates having the best scores or scores near the best scores.}
}
\value{
List of positions.
}
\description{
Identify tables which are the most likely to be the desired financial statements.
}
\author{
Nicolas Mangin
}
