% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_db_actions.R
\name{insert_clause}
\alias{insert_clause}
\title{Generate SQL insert statement}
\usage{
insert_clause(con, table, values = list(...), type = 1, cols = NULL, exec = T)
}
\arguments{
\item{table}{The name of the table to insert the values into. Must be quoted if the table name contains uppercase letters, spaces or special characters.}

\item{values}{A list of values to be inserted into the table.}

\item{type}{An integer indicating the type of insert statement to generate. Type 1 is a standard insert statement where all values are inserted into the table without specifying columns. Type 2 is an insert statement that specifies the columns to insert the values into.}

\item{cols}{A vector of column names to be specified in the insert statement. Only used if type = 2.}

\item{exec}{A logical value indicating whether or not the insert statement should be executed immediately. If set to TRUE, the insert statement will be executed using the RPostgreSQL::dbExecute function.}
}
\value{
A string containing the SQL insert statement.
}
\description{
Generates an SQL insert statement for a given table and values.
}
