% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_allocationclass.R
\name{internal_team_outsource_cf}
\alias{internal_team_outsource_cf}
\title{Perform account allocation to internal team for carried forward accounts based on outsource method.}
\usage{
internal_team_outsource_cf(id_col, dt, outsource_method = "equal", test = TRUE)
}
\arguments{
\item{dt}{The allocation data frame.}

\item{outsource_method}{The outsource method, either "equal" or "ratio".}

\item{test}{Logical indicating whether to run a test mode (default is TRUE).}
}
\value{
The reallocated data frame.
}
\description{
Perform account allocation to internal team for carried forward accounts based on outsource method.
}
