% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_allocationclass.R
\name{internal_team_outsource_new}
\alias{internal_team_outsource_new}
\title{Perform account allocation to internal team based on outsource method.}
\usage{
internal_team_outsource_new(
  id_col,
  agents,
  dt,
  outsource_method = "equal",
  ratio = NULL,
  test = TRUE
)
}
\arguments{
\item{agents}{A character vector of internal debt collectors' names.}

\item{dt}{The allocation data frame.}

\item{outsource_method}{The outsource method, either "equal" or "ratio".}

\item{ratio}{The allocation ratio for each agent (optional, required for "ratio" outsource method).}

\item{test}{Logical indicating whether to run a test mode (default is TRUE).}
}
\value{
The allocation data frame.
}
\description{
Perform account allocation to internal team based on outsource method.
}
