% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRR.R
\name{IRR}
\alias{IRR}
\title{Internal Rate of Return}
\usage{
IRR(CF = 0)
}
\arguments{
\item{CF}{Numeric vector. Set of periodic cash flows.}
}
\value{
Numeric. Periodic Internal Rate of Return associated with a series of period cash flows.
}
\description{
Wrapper emulating the MS Excel financial math function to compute the periodic Internal Rate of Return associated with a series of period cash flows.
}
\author{
Nicolas Mangin
}
