% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_settings.R
\name{is_endMonth}
\alias{is_endMonth}
\title{Check if a date is the last day of the month}
\usage{
is_endMonth(reportDate)
}
\arguments{
\item{reportDate}{A date in the format "YYYY-MM-DD" to be checked.}
}
\value{
Logical value: TRUE if the input date is the last day of the month, FALSE otherwise.
}
\description{
This function takes a date as input and checks if it is the last day of the month.
}
\keyword{date}
\keyword{day}
\keyword{last}
\keyword{month}
