% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_kami_excel}
\alias{load_kami_excel}
\title{Load data from all range of Excel sheet as character}
\usage{
load_kami_excel(path, sheet)
}
\arguments{
\item{path}{Path to the xls/xlsx file.}

\item{sheet}{Sheet to read. Either a string (the name of a sheet), or an
integer (the position of the sheet). Ignored if the sheet is specified via
\code{range}. If neither argument specifies the sheet, defaults to the first
sheet.}
}
\description{
Load data from all range of Excel sheet as character
}
