% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_data.R
\docType{data}
\name{market_data}
\alias{market_data}
\title{market_data}
\format{
A data frame with 93212 rows and 4 variables
\itemize{
\item  cik   character
\item  date   character
\item  adjusted_price   character
\item  dividend   character
}
}
\usage{
data(market_data)
}
\description{
market_data
}
\keyword{datasets}
