% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_settings.R
\name{passwd_generator}
\alias{passwd_generator}
\title{Generate a random password}
\usage{
passwd_generator(length = 8)
}
\arguments{
\item{length}{An integer specifying the length of the password. The default value is 8.}
}
\value{
A character string containing the random password.
}
\description{
This function generates a random password of a specified length. The password contains
alphanumeric characters and some special characters.
}
\examples{
passwd_generator(length = 12)
passwd_generator()
}
