% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{+.collectr}
\alias{+.collectr}
\title{Syntactic shortcut for adding an item to a collectr collection}
\usage{
\method{+}{collectr}(collection, item)
}
\arguments{
\item{collection}{A \code{collectr} object.}

\item{item}{A \code{collectr.item} or another object to add.}
}
\value{
A modified \code{collectr} collection with the new item.
}
\description{
Provides \code{+} as a shortcut for adding an item to a \code{collectr} collection.
}
