% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RATE.R
\name{RATE}
\alias{RATE}
\title{Periodic rate of return}
\usage{
RATE(NPER = 0, PMT = 0, PV = 0, FV = 0, type = 0)
}
\arguments{
\item{NPER}{Integer. Number of periods.}

\item{PMT}{Numeric. Cash flow at every period.}

\item{PV}{Numeric. Cash flow at the beginning of all the periods.}

\item{FV}{Numeric. Cash flow at the end of all the periods.}

\item{type}{Integer. 0 if payment at the end of the period or 1 if payment at the beginning of the period.}
}
\value{
Numeric.Periodic rate of return for an annuity.
}
\description{
Wrapper emulating the MS Excel financial math function to compute the periodic rate of return for an annuity.
}
\author{
Nicolas Mangin
}
