% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat_last.R
\name{repeat_last}
\alias{repeat_last}
\title{Replace NA by last known value}
\usage{
repeat_last(x, forward = TRUE, maxgap = Inf, na.rm = FALSE)
}
\arguments{
\item{x}{Vector.}

\item{forward}{Logical.}

\item{maxgap}{Numeric.}

\item{na.rm}{Logical.}
}
\value{
Vector
}
\description{
Replace NA in a vector by the last known value.
}
\author{
Nicolas Mangin
}
