% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_statements.R
\name{retrieve_statements}
\alias{retrieve_statements}
\title{Retrieve statements}
\usage{
retrieve_statements(tables, positions)
}
\arguments{
\item{tables}{List. all tables from a 10-K annual report.}

\item{positions}{List. Position of the tables containing financial statements}
}
\value{
List of formated financial statements.
}
\description{
Retrieve and format tables selected as financial statements.
}
\author{
Nicolas Mangin
}
