% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_statement_position.R
\name{return_statement_position}
\alias{return_statement_position}
\title{Find statements positions}
\usage{
return_statement_position(
  tables,
  patterns,
  remove = NA,
  minscore = 3,
  window = 2
)
}
\arguments{
\item{tables}{List. all tables from a 10-K annual report.}

\item{patterns}{List of character vectors. Patterns identifying the statements.}

\item{remove}{Numeric vector. Position of the tables to be removed.}

\item{minscore}{Integer. Minimum number of pattern occurrences to keep a table as a candidate for a statement.}

\item{window}{Integer. To keep candidates having the best scores or scores near the best scores.}
}
\value{
Numeric vector of positions
}
\description{
Identify the table which is the most likely to be the desired financial statement.
}
\author{
Nicolas Mangin
}
