% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_fs_patterns.R
\name{score_fs_patterns}
\alias{score_fs_patterns}
\title{Score patterns}
\usage{
score_fs_patterns(x, y)
}
\arguments{
\item{x}{Matrix. Table in which the terms should be counted.}

\item{y}{Character vector. Terms or expressions which should be counted in y.}
}
\value{
Integer. Count of occurrences of any term from y in x.
}
\description{
Count the occurrences of a term in a table to identify whether the table is one of the financial statements.
}
\author{
Nicolas Mangin
}
