% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shelve.R
\name{shelve}
\alias{shelve}
\title{Saves all the items in a collectr according to its options}
\usage{
shelve(object)
}
\arguments{
\item{object}{collectr object, currently only supports 'collectr.gg'}
}
\value{
invisible()
}
\description{
Saves all the items in a collectr according to its options
}
