% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clientprofile.R
\name{show_feedbackinput}
\alias{show_feedbackinput}
\title{Show Feedback Input Modal}
\usage{
show_feedbackinput(ns, username)
}
\arguments{
\item{username}{The username containing account information.}
}
\description{
This function displays a modal dialog for collecting customer feedback.
}
