% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statements.R
\docType{data}
\name{statements}
\alias{statements}
\title{statements}
\format{
A data frame with 45067 rows and 6 variables
\itemize{
\item  cik   character
\item  date   character
\item  reportDate   character
\item  id   character
\item  label   character
\item  amount   character
}
}
\usage{
data(statements)
}
\description{
statements
}
\keyword{datasets}
