% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_market_and_financial_data.R
\name{update_market_and_financial_data}
\alias{update_market_and_financial_data}
\title{Create Excel files for market and financial data}
\usage{
update_market_and_financial_data(type = "market", path = NA)
}
\arguments{
\item{type}{Character. Whether the function should create a "market" or "financial" dataset.}

\item{path}{Character. Path to the folder where the dataset should be saved.}
}
\value{
Excel files to be added to databases
}
\description{
Create Excel files for market and financial data based on companies in statements.
}
\author{
Nicolas Mangin
}
