% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_settings.R
\name{user_details_man}
\alias{user_details_man}
\title{Display a modal dialog for managing user details}
\usage{
user_details_man(ns, type = "new", ...)
}
\arguments{
\item{ns}{The namespace of the Shiny app.}

\item{type}{The type of user details to display ('new' or 'edit').}

\item{...}{Additional arguments to pass to the function.}
}
\value{
A Shiny modal dialog.
}
\description{
This function displays a modal dialog for adding or editing user details,
permissions, and password management settings.
}
