% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_settings.R
\name{user_permissions}
\alias{user_permissions}
\title{Create a user permissions form for a web application.}
\usage{
user_permissions(
  ns,
  group_auth,
  user_auth,
  selected_perms = NULL,
  fun = "edit_user_details"
)
}
\arguments{
\item{ns}{A string used to create unique names for the form inputs to avoid naming conflicts.}

\item{group_auth}{A vector of strings representing the modules or groups that the user can access.}

\item{user_auth}{A vector of strings representing the permissions that the user has for each module or group.}

\item{selected_perms}{A vector of strings representing the permissions that have been selected by the user.}

\item{fun}{A string representing the name of the function that will be called when the user clicks the "Save" button.}
}
\value{
A shiny UI object that can be used to render the user permissions form.
}
\description{
This function creates an interactive user permissions form for a web application using the Shiny package in R. The form allows the user to select their permissions for each module or group and displays the selected permissions in a separate input. The form includes a "Save" button that calls a specified function when clicked.
}
