% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect}
\alias{collect}
\title{Add an item to a collectr collection}
\usage{
collect(item, ..., collection = NULL)
}
\arguments{
\item{item}{a \code{collectr.item} to be added to a \code{collectr} or another R object to be converted into a \code{collectr.item}}

\item{...}{Named options for the new item's specific options.}

\item{collection}{A \code{collectr} object to add the item to. If NULL, a new \code{collectr.item} is created.}
}
\value{
A new \code{collectr.item} object or a modified \code{collectr} object with the new item included.
}
\description{
Either adds a \code{collectr.item} into a \code{collectr} or transforms an object that is not yet a \code{colelctr.item} into one.
}
