% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colrect.R
\name{colrect}
\alias{colrect}
\title{Collect clustered data on multiple sheets into rectangle dataframe}
\usage{
colrect(path, sheet, col)
}
\arguments{
\item{path}{Path to the xls/xlsx file.}

\item{sheet}{Sheet to read. Either a string (the name of a sheet), or an
integer (the position of the sheet). Ignored if the sheet is specified via
\code{range}. If neither argument specifies the sheet, defaults to the first
sheet.}

\item{col}{column position to search keyword}
}
\description{
Collect clustered data on multiple sheets into rectangle dataframe
}
