% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_market_data.R
\name{get_market_data}
\alias{get_market_data}
\title{Get data about stock price and dividends}
\usage{
get_market_data(filings)
}
\arguments{
\item{filings}{Tibble. Cik codes and date.}
}
\value{
Table with cik, date, adjusted_price and dividend.
}
\description{
Get data about stock price and dividends
}
\author{
Nicolas Mangin
}
