% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_market_data.R
\name{get_market_data}
\alias{get_market_data}
\title{Get data about stock price and dividends}
\usage{
get_market_data(cik = NA, ticker = NA, from = NA)
}
\arguments{
\item{cik}{Character. cik code of the company}

\item{ticker}{Character. Ticker of the company.}

\item{from}{Character. IDO code for the first date.}
}
\value{
Table with cik, date, adjusted_price and dividend.
}
\description{
Get data about stock price and dividends
}
\author{
Nicolas Mangin
}
